1. UserRepository【使用者資料】 JOB_STATUS【在職狀況】欄位 調整。 2. CommonRepository【系統共通功能邏輯】 GetRoles【角色列表】 新增 JOB_STATUS【在職狀況】欄位。 建置 PostCommonApprovers 【會簽常用名單】 API。

+++ OA_WEB_API/Models/BPMPro/UserModel.cs
@@ -256,7 +256,7 @@ namespace OA_WEB_API.Models
         public int JOB_GRADE { get; set; }
 
         /// <summary>工作狀態</summary>
-        public int JOB_STATUS { get; set; }
+        public Byte JOB_STATUS { get; set; }
 
         /// <summary>是否為正職部門</summary>
         public int IS_MAIN_JOB { get; set; }
